perception: fix sensor fusion code stype

+++ modules/perception/obstacle/fusion/probabilistic_fusion/pbf_sensor.cc
@@ -16,6 +16,9 @@
 
 #include "modules/perception/obstacle/fusion/probabilistic_fusion/pbf_sensor.h"
 
+#include <string>
+#include <vector>
+
 namespace apollo {
 namespace perception {
 
@@ -63,7 +66,7 @@ void PbfSensor::AddFrame(const SensorObjects &frame) {
   pbf_frame->seq_num = frame.seq_num;
 
   pbf_frame->objects.resize(frame.objects.size());
-  for (int i = 0; i < (int)frame.objects.size(); i++) {
+  for (size_t i = 0; i < frame.objects.size(); i++) {
     PbfSensorObjectPtr obj(new PbfSensorObject());
     obj->timestamp = frame.timestamp;
     obj->sensor_type = frame.sensor_type;
@@ -71,7 +74,7 @@ void PbfSensor::AddFrame(const SensorObjects &frame) {
     obj->sensor_id = GetSensorType(frame.sensor_type);
     pbf_frame->objects[i] = obj;
   }
-  if ((int)(frames_.size()) > s_max_cached_frame_number_) {
+  if (frames_.size() > s_max_cached_frame_number_) {
     frames_.pop_front();
   }
   frames_.push_back(pbf_frame);
@@ -83,7 +86,7 @@ bool PbfSensor::GetPose(double time_stamp, Eigen::Matrix4d *pose) {
     return false;
   }
 
-  for (int i = (int)frames_.size() - 1; i >= 0; i--) {
+  for (int i = frames_.size() - 1; i >= 0; i--) {
     double time_diff = time_stamp - frames_[i]->timestamp;
     if (fabs(time_diff) < 1.0e-3) {
       *pose = frames_[i]->sensor2world_pose;
